requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
markdown==3.5.1
python-slugify==8.0.1
Werkzeug==2.3.7 